# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bidly API Documentation",
        "description": "행사 기획 대행사와 용역사를 연결하는 입찰 플랫폼 API",
        "contact": {
            "name": "Bidly Support Team",
            "email": "support@bidly.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/announcements/{announcement}/proposals": {
            "post": {
                "tags": [
                    "Proposal Management"
                ],
                "summary": "제안서 제출",
                "description": "특정 공고에 제안서를 제출합니다. 용역사만 제출 가능하며, 공고당 한 번만 제출할 수 있습니다.",
                "operationId": "cfbb9c13450a9c08c28038e16335fc09",
                "parameters": [
                    {
                        "name": "announcement",
                        "in": "path",
                        "description": "공고 ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "proposed_price": {
                                        "description": "제안 금액",
                                        "type": "number",
                                        "example": 42000000
                                    },
                                    "proposal_text": {
                                        "description": "제안서 내용",
                                        "type": "string",
                                        "example": "저희 회사는 10년간의 무대 설치 경험을 바탕으로..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "제안서 제출 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "제안서가 성공적으로 제출되었습니다."
                                        },
                                        "proposal": {
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "proposed_price": {
                                                    "type": "number"
                                                },
                                                "proposal_text": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "submitted"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "권한 없음",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "제안서 제출 권한이 없습니다."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "중복 제출",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "이미 해당 공고에 제안서를 제출했습니다."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}